Format function argument separator comments
 ## Summary

This is a complete rewrite of the handling of `/` and `*` comment handling in function signatures. The key problem is that slash and star don't have a note. We now parse out the positions of slash and star and their respective preceding and following note. I've left code comments for each possible case of function signature structure and comment placement

## Test Plan

I extended the function statement fixtures with cases that i found. If you have more weird edge cases your input would be appreciated.