Add test case for false positive
But this commit doesn't actually fix the problem.

The problem is that the try statements aren't handled yet and simple
continue with the next block, which in the case of a `while True` loop
creates an infinite loop. Thus the rule triggers on any statements after
the while loop, but this is incorrect.