Remove tag information from RustPython-Parser dependency
Following https://github.com/astral-sh/RustPython-Parser/pull/27 we now cherry-pick commits onto our fork instead of rebasing our fork on top of the upstream which means we do not overwrite history and a tag is not necessary to preserve the pinned commit.

In the future, we may rewrite the history in our fork. If we do, we should return to tagging the commits.